Beacon spotting and UI updates
Many updates to add inline spotting of beacons
Messages cent as CQ to the group now attach themselves to the correct user if a callsign is at the front of the message
clicking the remote stations location in chat window shows them on map, might add a button for this.
only one map user "infowindow" at any time.